[AN] feat(#267): api 클래스 분리
